Fixing the split cluster issue caused by chunked query processing.
